Put intent/event handlers in their own file
